refactor: realised that we will obviously need to have float support

+++ matrix/matrix.go
@@ -1,6 +1,6 @@
 package matrix
 
-type matrix = [][]int
+type matrix = [][]float64
 
 func Transpose(m matrix) matrix {
 	m2 := make(matrix, len(m), len(m))
